requirements.txt
alembic==1.0.7
appdirs==1.4.3
astroid==2.1.0
attrs==18.2.0
black==18.9b0
boto3==1.9.99
botocore==1.12.99
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
decorator==4.3.2
docutils==0.14
idna==2.8
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
lxml==4.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
nose2==0.8.0
numpy==1.16.1
packaging==19.0
pandas==0.24.1
pbr==5.1.2
pip-review==1.0
psycopg2-binary==2.7.7
pydocstyle==3.0.0
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
rope==0.12.0
s3transfer==0.2.0
selenium==3.141.0
six==1.12.0
snowballstemmer==1.2.1
SQLAlchemy==1.2.18
sqlalchemy-migrate==0.12.0
sqlparse==0.2.4
Tempita==0.5.2
toml==0.10.0
tqdm==4.31.1
typed-ast==1.3.1
tzlocal==1.5.1
urllib3==1.24.1
w3lib==1.20.0
wrapt==1.11.1
